+++ src/EditableTextLabel.js
@@ -1,20 +1,34 @@
+// @flow
 import React, { PureComponent } from 'react';
 import PropTypes from 'prop-types';
+import keycode from 'keycode';
+
+import type { ReactChildren } from 'react-flow-types';
 
 import { getTextLayoutProps } from './mixins/rowComp';
 
-import EditableText, { PureEditableText } from './EditableText';
+import EditableText from './EditableText';
 import Icon from './Icon';
 import TextLabel from './TextLabel';
 
 import { STATUS_CODE as STATUS } from './StatusIcon';
 
-class EditableTextLabel extends PureComponent {
+export type Props = {
+    inEdit: boolean,
+    onEditRequest: () => void,
+    onEditEnd: (payload?: { value: string | null, event: Event }) => void,
+    // #FIXME: use exported Flow types
+    icon?: string,
+    basic?: ReactChildren,
+    align?: string,
+    status?: string | null,
+};
+
+class EditableTextLabel extends PureComponent<Props, Props, any> {
     static propTypes = {
         inEdit: PropTypes.bool,
         onEditRequest: PropTypes.func,
-        // <EditableText> callbacks
-        onEditEnd: PureEditableText.propTypes.onEditEnd,
+        onEditEnd: PropTypes.func,
         // <TextLabel> props
         icon: TextLabel.propTypes.icon,
         basic: TextLabel.propTypes.basic,
@@ -25,18 +39,13 @@ class EditableTextLabel extends PureComponent {
     static defaultProps = {
         inEdit: false,
         onEditRequest: () => {},
-        onEditEnd: PureEditableText.defaultProps.onEditEnd,
+        onEditEnd: () => {},
+        // <TextLabel> props
         icon: TextLabel.defaultProps.icon,
         basic: TextLabel.defaultProps.basic,
         align: TextLabel.defaultProps.align,
         status: TextLabel.defaultProps.status,
     };
-
-    componentDidUpdate() {
-        if (this.props.inEdit) {
-            this.editableTextRef.getRenderedComponent().focusInputNode();
-        }
-    }
 
     handleDoubleClick = () => {
         /**
@@ -52,6 +61,30 @@ class EditableTextLabel extends PureComponent {
          * We should rely on visible buttons or menus to trigger edit.
          */
         this.props.onEditRequest();
+    }
+
+    handleInputBlur = (event: Event & { currentTarget: HTMLInputElement }) => {
+        this.props.onEditEnd({
+            value: event.currentTarget.value,
+            event,
+        });
+    }
+
+    handleInputKeyDown = (event: KeyboardEvent & { currentTarget: HTMLInputElement }) => {
+        switch (event.keyCode) {
+            case keycode('Enter'):
+                // Blur the input, and trigger `onEditEnd` in blur handler
+                event.currentTarget.blur();
+                break;
+            case keycode('Escape'):
+                this.props.onEditEnd({
+                    value: null,
+                    event,
+                });
+                break;
+            default:
+                break;
+        }
     }
 
     render() {
@@ -77,9 +110,12 @@ class EditableTextLabel extends PureComponent {
             <TextLabel {...labelProps}>
                 {icon && <Icon type={icon} />}
                 <EditableText
-                    ref={(ref) => { this.editableTextRef = ref; }}
                     defaultValue={basic}
-                    onEditEnd={onEditEnd}
+                    onBlur={this.handleInputBlur}
+                    input={{
+                        autoFocus: true,
+                        onKeyDown: this.handleInputKeyDown,
+                    }}
                     {...layoutProps} />
             </TextLabel>
         );