+++ src/__tests__/EditableText.test.js
@@ -1,14 +1,10 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import { shallow, mount } from 'enzyme';
-import keycode from 'keycode';
+import { shallow } from 'enzyme';
 
+import EditableBasicRow from '../EditableBasicRow';
 import EditableText, { PureEditableText } from '../EditableText';
 import { PureText } from '../Text';
-
-function delayForMilliseconds(millisecond) {
-    return new Promise(resolve => setTimeout(resolve, millisecond));
-}
 
 describe('withStatus(EditableText)', () => {
     it('renders without crashing', () => {
@@ -20,172 +16,86 @@ describe('withStatus(EditableText)', () => {
 });
 
 describe('pure <PureEditableText>', () => {
-    it('renders an <input type="text"> inside <BasicRow>', () => {
-        const wrapper = mount(<PureEditableText />);
+    it('renders <EditableBasicRow> with pure <Text>', () => {
+        const wrapper = shallow(<PureEditableText />);
 
-        expect(
-            wrapper.find('BasicRow').containsMatchingElement(<input type="text" />)
-        ).toBeTruthy();
+        expect(wrapper.find(PureText)).toHaveLength(1);
+        expect(wrapper.find(PureText).dive().find(EditableBasicRow)).toHaveLength(1);
     });
 
-    it('renders basic label with placeholder if <input> has no value', () => {
-        const wrapper = mount(<PureEditableText placeholder="FooBar Placeholder" />);
-        expect(wrapper.find('BasicRow').text()).toBe('FooBar Placeholder');
-    });
+    it('updates state on <EditableBasicRow> focus/blur', () => {
+        const wrapper = shallow(<PureEditableText />);
+        const rowWrapper = wrapper.find(PureText).dive().find(EditableBasicRow);
 
-    it('renders basic label with the value from <input>', () => {
-        const wrapper = mount(<PureEditableText value="Foo" />);
-        expect(wrapper.find('BasicRow').text()).toBe('Foo');
-
-        wrapper.setProps({ value: 'Bar' });
-        expect(wrapper.find('BasicRow').text()).toBe('Bar');
-    });
-
-    it('enters focused mode when <input> is focused.', () => {
-        const wrapper = mount(<PureEditableText />);
         expect(wrapper.state('focused')).toBeFalsy();
-        expect(wrapper.hasClass('gyp-editable-text--focused')).toBeFalsy();
 
-        wrapper.find('input').simulate('focus');
+        rowWrapper.simulate('focus');
         expect(wrapper.state('focused')).toBeTruthy();
-        expect(wrapper.hasClass('gyp-editable-text--focused')).toBeTruthy();
 
-        wrapper.find('input').simulate('blur');
+        rowWrapper.simulate('blur');
         expect(wrapper.state('focused')).toBeFalsy();
-        expect(wrapper.hasClass('gyp-editable-text--focused')).toBeFalsy();
     });
 
-    it('renders without status elements in focused mode', () => {
-        const wrapper = shallow(
-            <PureEditableText
-                statusIcon={<span data-icon />}
-                errorMsg="Foo-Bar" />
-        );
+    it('forwards onFocus/onBlur events', () => {
+        const handleFocus = jest.fn();
+        const handleBlur = jest.fn();
 
-        expect(wrapper.prop('statusIcon')).toEqual(<span data-icon />);
+        const wrapper = shallow(<PureEditableText onFocus={handleFocus} onBlur={handleBlur} />);
+        const rowWrapper = wrapper.find(PureText).dive().find(EditableBasicRow);
+
+        expect(handleFocus).not.toHaveBeenCalled();
+        expect(handleBlur).not.toHaveBeenCalled();
+
+        rowWrapper.simulate('focus');
+        expect(handleFocus).toHaveBeenCalledTimes(1);
+        expect(handleBlur).not.toHaveBeenCalled();
+
+        rowWrapper.simulate('blur');
+        expect(handleFocus).toHaveBeenCalledTimes(1);
+        expect(handleBlur).toHaveBeenCalledTimes(1);
+    });
+
+    it('strips status props from <PureText> when focused', () => {
+        const mockedIcon = <span data-icon />;
+        const wrapper = shallow(<PureEditableText statusIcon={mockedIcon} errorMsg="Foo-Bar" />);
+
+        expect(wrapper.prop('statusIcon')).toEqual(mockedIcon);
         expect(wrapper.prop('errorMsg')).toBe('Foo-Bar');
 
         wrapper.setState({ focused: true });
-        expect(wrapper.prop('statusIcon')).toBe(PureText.defaultProps.statusIcon);
-        expect(wrapper.prop('errorMsg')).toBe(PureText.defaultProps.errorMsg);
+
+        expect(wrapper.prop('statusIcon')).toBeUndefined();
+        expect(wrapper.prop('errorMsg')).toBeUndefined();
     });
 
-    it('works like a controlled input when given "value" prop', () => {
-        const wrapper = mount(<PureEditableText value="Foo" />);
-        const input = wrapper.find('input');
+    it('passes status to <EditableBasicRow> and sets it as readOnly when status is loading', () => {
+        const wrapper = shallow(<PureEditableText />);
 
-        expect(input.prop('value')).toBe('Foo');
+        let rowWrapper = wrapper.find(PureText).dive().find(EditableBasicRow);
+        expect(rowWrapper.prop('readOnly')).toBeFalsy();
 
-        input.simulate('change', {
-            target: { value: 'Bar' },
-        });
-        expect(input.prop('value')).toBe('Foo');
+        wrapper.setProps({ status: 'loading' });
+        // Refreshes row from updated wrapper
+        rowWrapper = wrapper.find(PureText).dive().find(EditableBasicRow);
+        expect(rowWrapper.prop('readOnly')).toBeTruthy();
+        expect(rowWrapper.prop('status')).toBe('loading');
+
+        wrapper.setProps({ status: 'success' });
+        // Refreshes row from updated wrapper
+        rowWrapper = wrapper.find(PureText).dive().find(EditableBasicRow);
+        expect(rowWrapper.prop('readOnly')).toBeFalsy();
+        expect(rowWrapper.prop('status')).toBe('success');
     });
 
-    it('works like an uncontrolled input when not given "value" prop', () => {
-        const wrapper = mount(<PureEditableText />);
-        const input = wrapper.find('input');
+    it('passes unknown props to <EditableBasicRow>', () => {
+        const wrapper = shallow(<PureEditableText noGrow foo bar="Bar" />);
+        const rowWrapper = wrapper.find(PureText).dive().find(EditableBasicRow);
 
-        expect(input.prop('value')).toBe('');
+        // Known prop for <PureText>
+        expect(rowWrapper.prop('noGrow')).toBeUndefined();
 
-        input.simulate('change', {
-            target: { value: 'Bar' },
-        });
-        expect(input.prop('value')).toBe('Bar');
-    });
-
-    it('does not clear cached input value when goes from controlled input to uncontrolled', () => {
-        const wrapper = mount(<PureEditableText value="Foo" />);
-        const input = wrapper.find('input');
-
-        expect(input.prop('value')).toBe('Foo');
-
-        wrapper.setProps({ value: null });
-        expect(input.prop('value')).toBe('Foo');
-    });
-
-    it('can have a default value', () => {
-        const wrapper = mount(<PureEditableText defaultValue="Foo" />);
-        const input = wrapper.find('input');
-
-        expect(input.prop('value')).toBe('Foo');
-
-        input.simulate('change', {
-            target: { value: 'Bar' },
-        });
-        expect(input.prop('value')).toBe('Bar');
-    });
-
-    it('notifies edit end on input blur', () => {
-        const handleEditEnd = jest.fn();
-        const wrapper = mount(<PureEditableText onEditEnd={handleEditEnd} />);
-        const input = wrapper.find('input');
-
-        input.simulate('blur', { target: { value: 'Foo' } });
-        expect(handleEditEnd).toHaveBeenLastCalledWith({ value: 'Foo', reset: false });
-    });
-
-    it('notifies edit end on Enter key', () => {
-        const handleEditEnd = jest.fn();
-        const wrapper = mount(<PureEditableText onEditEnd={handleEditEnd} />);
-        const input = wrapper.find('input');
-
-        input.simulate('keydown', {
-            target: { value: 'Bar' },
-            keyCode: keycode('Enter'),
-        });
-        expect(handleEditEnd).toHaveBeenLastCalledWith({ value: 'Bar', reset: false });
-        expect(handleEditEnd).toHaveBeenCalledTimes(1);
-    });
-
-    it('notifies edit end and reminds to reset on Escape key', () => {
-        const handleEditEnd = jest.fn();
-        const wrapper = mount(<PureEditableText onEditEnd={handleEditEnd} />);
-        const input = wrapper.find('input');
-
-        input.simulate('keydown', {
-            target: { value: 'Bar' },
-            keyCode: keycode('Escape'),
-        });
-        expect(handleEditEnd).toHaveBeenLastCalledWith({ value: 'Bar', reset: true });
-        expect(handleEditEnd).toHaveBeenCalledTimes(1);
-    });
-
-    it('prevents another notifiying edit end within short time interval', async () => {
-        const handleEditEnd = jest.fn();
-        const wrapper = mount(<PureEditableText onEditEnd={handleEditEnd} />);
-        const input = wrapper.find('input');
-
-        input.simulate('keydown', {
-            keyCode: keycode('Escape'),
-        });
-        expect(handleEditEnd).toHaveBeenCalledTimes(1);
-
-        await delayForMilliseconds(100);
-
-        input.simulate('blur');
-        expect(handleEditEnd).toHaveBeenCalledTimes(1);
-    });
-
-    it('does not interupt user input on non-Enter/Esc keys', () => {
-        const handleEditEnd = jest.fn();
-        const wrapper = mount(<PureEditableText onEditEnd={handleEditEnd} />);
-        const input = wrapper.find('input');
-
-        input.simulate('keydown', {
-            target: { value: 'Foo' },
-            keyCode: keycode('B'),
-        });
-        expect(handleEditEnd).not.toHaveBeenCalled();
-    });
-
-    it('can manually focus its <input> on DOM', () => {
-        const wrapper = mount(<PureEditableText />);
-        const input = wrapper.instance().inputNode;
-
-        const spy = jest.spyOn(input, 'focus');
-        wrapper.instance().focusInputNode();
-
-        expect(spy).toHaveBeenCalled();
+        // Unknown props
+        expect(rowWrapper.prop('foo')).toBeTruthy();
+        expect(rowWrapper.prop('bar')).toBe('Bar');
     });
 });